requirements.txt
jupyter
pandas
opencv-python
matplotlib
SimpleITK
imutils
pydicom
nibabel
scikit-learn
vtk
pydicom_seg
pyntcloud
scikit-image
trimesh
medpy
torchinfo
albumentations
torch
torchvision
torchaudion
tensorflow-gpu

